refactor to updated harp core

+++ firmware/inc/white_rabbit_app.h
@@ -34,7 +34,7 @@ const size_t REG_COUNT{5};
 
 // pre-computed value for when to emit periodic counter msgs.
 extern uint32_t counter_interval_us;
-extern uint32_t last_msg_emit_time_us;
+extern uint64_t last_msg_emit_time_us;
 extern bool was_synced;
 
 #pragma pack(push, 1)
